// src/types/api-types/utils/Itslearning.RestApi.Entities.CourseCard.ts
export type ItslearningRestApiEntitiesCourseCard = {
  CourseId: number;
  FriendlyName: any;
  Title: string;
  TeacherName: any;
  TeacherPictureUrl: any;
  NumberOfTeachers: number;
  Teachers: any[];
  CourseColorClass: string;
  NumberOfAnnouncements: number;
  NumberOfTasks: number;
  NumberOfFollowUpTasks: number;
  LastUpdatedDisplayTime: string;
  LastUpdated: string;
  LastOnlineDisplayTime: string;
  LastOnline: string;
  IsFavouriteCourse: boolean;
  CanAdminCourse: boolean;
};